pyproject.toml
[project]
name = "crawl-sheet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langfuse>=3.0.0",
    "openinference-instrumentation-smolagents>=0.1.12",
    "opentelemetry-exporter-otlp>=1.34.0",
    "opentelemetry-sdk>=1.34.0",
    "requests>=2.32.3",
    "scrapy>=2.13.1",
    "smolagents[gradio,mcp,openai,telemetry,toolkit]>=1.17.0",
]
